# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bidirectional 837 Translator",
    "version": "v1",
    "description": "Governed API surface for the bi-directional 837P generation and translation engine. This document is the authored contract required by governance.txt Section 4 ('API Contracts First'). It is committed before the controllers beneath it exist, and Translator.Api.Tests asserts that the running application conforms to it. Routes not named by governance are recorded in ADR-009."
  },
  "paths": {
...
    "/api/v1/codes": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "ListMedicalCodes",
        "summary": "List the catalogued medical codes, their category and their standard charge.",
        "description": "Feeds the generation form's code selector. Every code is HCPCS Level II, is priced by a published CMS fee schedule, and carries the category the generator groups it under.",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Restrict the result to one category.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalogue.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalCodeDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such category is catalogued.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MedicalCodeDto": {
        "type": "object",
        "required": [
          "Code",
          "Category",
          "Description",
          "StandardCharge"
        ],
        "properties": {
          "Code": {
            "type": "string",
            "maxLength": 5
          },
          "Category": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          },
          "StandardCharge": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}